Disable outside click & extract to hooks dir.
